[Level 1 Piece of Cake]: Move test classes into domain class
