+++ domain/build.gradle
@@ -43,5 +43,8 @@ dependencies {
 
     testImplementation libs.test.junit4
     testImplementation libs.hilt.compiler
+    testImplementation libs.mockito.kotlin
+    testImplementation libs.kotlinx.coroutines.test
+    testImplementation libs.arch.core.testing
     kaptTest libs.hilt.compiler
 }
