+++ domain/src/test/java/com/roque/rueda/gbm/coding/domain/base/TestCoroutineRule.kt
@@ -1,0 +1,28 @@
+package com.roque.rueda.gbm.coding.domain.base
+
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.test.*
+import org.junit.rules.TestRule
+import org.junit.runner.Description
+import org.junit.runners.model.Statement
+
+@ExperimentalCoroutinesApi
+class TestCoroutineRule : TestRule {
+
+    private val testCoroutineDispatcher = StandardTestDispatcher()
+    private val testCoroutineScope = TestScope()
+
+    override fun apply(base: Statement, description: Description): Statement = object : Statement() {
+        override fun evaluate() {
+            Dispatchers.setMain(testCoroutineDispatcher)
+
+            base.evaluate()
+
+            Dispatchers.resetMain()
+        }
+    }
+
+    fun runBlockingTest(block: suspend TestScope.() -> Unit) =
+        testCoroutineScope.runTest { block() }
+}
